+++ lib/components/menu_bar/README.md
@@ -8,42 +8,35 @@ Demo video (5 minutes): https://youtu.be/k1kiCL9oMf4
 ## How to use the MenuBar component in your Scenic app
 
 This is an example of all the code required to render a MenuBar.
+Setting `pin: {0, 0}` places the MenuBar in the top-left corner.
 
 ```
-vp_width = 1000         # fetch the viewport width from Scenic
-menu_bar_height = 40    # pick whatever you like
-{:fixed, 220}           # sets how wide the columns of the menus are
-
-# calculate font data
-{:ok, ibm_plex_mono_metrics} =
-    TruetypeMetrics.load("./assets/fonts/IBMPlexMono-Regular.ttf")
-
-font = %{               # Pass in font metrics data for the MenuBar
-    size: 36,           # This is the dize of the font for the main MenuBar (not sub-menus)
-    metrics: ibm_plex_mono_metrics
-}
-
-sub_menu_options = %{
-    height: 40,         # the block-height of sub-menu item rectangles
-    font_size: 22       # font-size to use in the sub-menus
-}
-
-menu_map = calc_menu_map() # see below for more explanation about this
+vp_width = 800 # need to pass in the ViewPort width
 
 Scenic.Graph.build()
 |> ScenicWidgets.MenuBar.add_to_graph( %{
         frame: ScenicWidgets.Core.Structs.Frame.new(
             pin: {0, 0},
-            size: {vp_width, menu_bar_height}),
-        menu_map: menu_map,
-        item_width: {:fixed, 180},
-        font: menubar_font,
-        sub_menu: menu_bar.sub_menu
+            size: {vp_width, _menu_bar_height = 60}),
+        menu_map: [
+            {:sub_menu, "Ice Cream", [
+                {"Chocolate", fn -> IO.puts "clicked: `Chocolate`!" end},
+                {"Vanilla", fn -> IO.puts "clicked: `Vanilla`!" end}
+            ]},
+            {:sub_menu, "Ninja Turtles", [
+                {"Leonardo", fn -> IO.puts "clicked: `Leonardo`!" end},
+                {"Raphael", fn -> IO.puts "clicked: `Raphael`!" end},
+                {"Donatello", fn -> IO.puts "clicked: `Donatello`!" end},
+                {"Michelangelo", fn -> IO.puts "clicked: `Michelangelo`!" end},
+            ]}
+        ]
 })
 ```
 
 `ScenicWidgets.Core.Structs.Frame` is a struct also defined inside
 ScenicContrib, it is just a fancy definition for a rectangular box.
+
+### Defining the MenuMap
 
 The actual contents of the MenuBar is completely customizable, but
 you need to pass in a specifically shaped tree (made out of lists).
@@ -82,11 +75,12 @@ end
 ```
 
 Note that each actual item (that isn't a sub-menu) is defined by
-a tuple which looks like `{"item label", function}`. This function
+a tuple which looks like `{"item label", function/0}`. This function
 will be executed when the menu item gets clicked, and can be anything
 you like, although there's no way to dynamically send it arguments
 at this time, so it must have an arity of zero (but you could go and
-fetch data from somewhere else if you so wished.)
+fetch data from somewhere else if you so wished, just wrap that
+function inside an arity-zero one, e.g. `fn -> do_whatever(x) end`)
 
 Note that functions MUST have side-effects to even work... this is
 because the function gets executed in the context of the MenuBar process,
@@ -109,7 +103,7 @@ is supposed to react to the button click, e.g.
 ```
 
 Using `IO.puts` just ensures that a message shows up in the IEx console
-and is not necessary, sometimes useful.
+and is not necessary, just sometimes useful.
 
 To call a zero-arity defined in a different module, simply use
 the function-capture syntax native to Elixir, e.g.
@@ -117,6 +111,83 @@ the function-capture syntax native to Elixir, e.g.
 ```
 {"Strawberry", &IceCream.Flavour.strawberry/0}
 ```
+
+## Dynamically changing the menu-map
+
+The MenuBar can have it's contents updated at any time - this is expecially
+useful for making context-aware menus, e.g. if one of your sub-menus showed
+a list of open files (as is the case for [Flamelex](https://github.com/JediLuke/flamelex)) then opening
+a new file should update the menu-bar to show this new file (which Flamelex does!)
+
+To update the menu map, simply cast to the component with a new MenuMap.
+
+```
+# the Component is automatically registered with this name
+GenServer.cast(ScenicWidgets.MenuBar, {:put_menu_map, new_menu_map})
+
+# alternatively, if you render this component from within another
+# scenic component, you can use `cast_children/2`
+cast_children(scene, {:put_menu_map, new_menu_map})
+```
+
+As hinted at above by the `calc_menu_map()` function, you can
+compute the MenuMap tree based on whatever state you like, and use
+this method to compute new Menu mappings whenever you want to change
+them, based on whatever your application is doing.
+
+### Optional args
+
+There are a number of customizations to the MenuBar which are possible
+by passing them in as arguments when creating the Graph.
+
+Here is an example of using all of the extended options
+
+```
+vp_width = 1000         # fetch the viewport width from Scenic
+menu_bar_height = 40    # pick whatever you like
+{:fixed, 220}           # sets how wide the columns of the menus are
+
+{:ok, ibm_plex_mono_metrics} =
+    TruetypeMetrics.load("./assets/fonts/IBMPlexMono-Regular.ttf")
+
+font = %{
+    name: :ibm_plex_mono,   # pass in the custom font here
+    size: 36,               # This is the size of the font for the main MenuBar (not sub-menus)
+    metrics: ibm_plex_mono_metrics      # pass in the FontMetrics we calculated above
+}
+
+sub_menu_options = %{
+    height: 40,         # the block-height of sub-menu item rectangles
+    font_size: 22       # font-size to use in the sub-menus
+}
+
+Scenic.Graph.build()
+|> ScenicWidgets.MenuBar.add_to_graph( %{
+        frame: ScenicWidgets.Core.Structs.Frame.new(
+            pin: {0, 0},
+            size: {vp_width, menu_bar_height}),
+        menu_map: menu_map,
+        item_width: {:fixed, 180},
+        font: menubar_font,
+        sub_menu: menu_bar.sub_menu
+})
+```
+
+One shortcut, if you just want to use a custom font but are happy to
+keep the default size, is to pass in the font's name (as an atom):
+
+```
+Scenic.Graph.build()
+|> ScenicWidgets.MenuBar.add_to_graph( %{
+        frame: ScenicWidgets.Core.Structs.Frame.new(
+            pin: {0, 0},
+            size: {vp_width, menu_bar_height}),
+        menu_map: menu_map(),
+        item_width: {:fixed, 180},
+        font: :ibm_plex_mono
+})
+```
+
 
 ### MenuBar controls
 
@@ -130,23 +201,6 @@ Right now the MenuBar doesn't recognise when you move the mouse horizontally
 away from a sub-menu, so it stays open. This is technically a bug, but
 surprisingly it is no issue in practice (at least for me).
 
-## Dynamically changing the menu-map
-
-The MenuBar can have it's contents updated at any time - this is expecially
-useful for making context-aware menus, e.g. if one of your sub-menus showed
-a list of open files (as is the case for [Flamelex](https://github.com/JediLuke/flamelex)) then opening
-a new file should update the menu-bar to show this new file (which it does!)
-
-To update the menu map, simply cast to the component with a new MenuMap.
-
-```
-# the Component is automatically registered with this name
-GenServer.cast(ScenicWidgets.MenuBar, {:put_menu_map, new_menu_map})
-
-# alternatively, if you render this component from within another
-# scenic component, you can use `cast_children/2`
-cast_children(scene, {:put_menu_map, new_menu_map})
-```
 
 ## Bonus - a trick to show arity/0 functions in the menu
